use threads when safe
libnih, when not built with --enable-threading, cannot be safely
used by a threaded application.  Detect whether it is built to be
threadsafe using a new libnih helper, and, if so, run threaded by
(a) not passing '-s' to fuse, and (b) making the dbus connection
and detected api version thread-local.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>