swap out libnih and libnih-dbus for glib
The motivation is to make threading possible, to hopefully greatly
speed up systemd startup inside containers.

This required converting all the nih-ified memory tracking.  (Some
of this can probably be done smarter in a more glib-friendly way, i.e.
avoiding doing a glib string alloc followed by strdup followed by
freeing the glib string)

We open a single dbus connection for all threads to use.  If that
connection is closed (i.e. cgmanager exits / restarts) the first
task to find it so takes a mutex and attempts to reconnect, once
per second, until it is reconnected.

When creating a directory for non-root user, execute a new binary
to get a clean dbus session as that user.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>