don't let idletime be > reaperage
This is not a good way to do this.  We should decide on a proper
heuristic.  We could take something like reaperage * (idletime/total_uptime),
but that doesn't scale for how much our own container used the cpu nor
for time.

I will open a github issue to fix this.

However as it currently stands the test_proc testcase was failing, this
at least lets it pass.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>