Use clone instead of fork for PID translation
Because of the different signatures of fork() and clone(),
pid_to_ns and pid_from_ns get an additional wrapper that is
passed to clone(). To pass the needed arguments to
pid_ns_clone_wrapper, a new struct called pid_ns_clone_args
is introduced.

The return type of pid_to_ns and pid_from_ns need to be
changed to int, returning equals exiting with clone().

(serge - inline fix of erorr typo which bled through from the original)

Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>
Acked-by: Serge E. Hallyn <serge.hallyn@ubuntu.com>