pam_cgfs: deal with hybrid cgroupfs mounts
When v1 hierarchies are mixed with a v2 hierarchy we currently crash since we
do not handle that case. This adds logic to at least place us in writeable
cgroups and to keep us functional. For now, we simply ignore the v2 hierarchy.
We will add specific logic to deal with it after we know what certain init
systems are planning wrt to v1 vs. v2 cgroup hierarchies.

Signed-off-by: Christian Brauner <christian.brauner@canonical.com>