Fix build on ia64
The relevant code was added in commit 35acc24, but the function/macro
prctl_arg() didn't seem to be defined anywhere in the repo. lxc
currently has a corresponding macro defined in src/lxc/macro.h that
casts the value to an unsigned long. But 0 doesn't require any special
handling, so remove the call to prctl_arg().

Verified that the code compiles properly on Debian's ia64 porterbox
(yttrium).

Signed-off-by: Mathias Gibbens <gibmat@debian.org>