lxcfs/bindings: add a flag for overriding the runtime dir
Adds a --runtime-dir cli flag which overrides the /run dir in the lxcfslib.
This ended up being kind of tricky because of how lxcfslib can be reloaded and
its use of a library constructor.

In order read the cli flag and then set a variable in the library, I removed
the contstructor and made init happen as part of the fuse load/reload.

I also added the runtime field to the lxcfs_opts struct and upped its version
for backwards compatibility.

Signed-off-by: Sebastien Dabdoub <sebastien.dabdoub@gmail.com>
- permission mask change 0755 -> 0700
- prevent potential NULL-pointer dereference in lxcfs_fuse_init()
- commit message edits
- one commit was squashed
Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>