src/bindings: add keep_on_reload field to struct pidns_store
Let's add keep_on_reload field to struct pidns_store.
The idea behind it is that if this flag is set to true, then
pidns_store entry won't be considered as a cache item which can
be dropped. But instead, it will be kept across liblxcfs reloads
and droped only if a pid namespace it refers die.

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>